ext4: limit group search loop for non-extent files
Patchwork-id: 58796
O-Subject: [PATCH RHEL6.5 2/2] ext4: limit group search loop for non-extent files
Bugzilla: 958939
RH-Acked-by: Lachlan McIlroy <lmcilroy@redhat.com>

For:

Bug 958939 - RHEL6.3: kernel panic during fsync, RIP ext4_mb_good_group,
after online resize of LUN on SAN from 500 to 800 GB

Backport of the following commit (currently only in the ext4
git tree, not yet in Linus' tree).

With patch 1/2 in this series, this isn't strictly needed, because
in RHEL6 we will never grow past 16T, which is the non-extent-file
limit.  So as long as we update s_blockfile_groups properly during
resize, the ">" condition will never be true.

But having this is good for completeness.

Thanks,
-Eric

>From e6155736ad76b2070652745f9e54cdea3f0d8567 Mon Sep 17 00:00:00 2001
From: Lachlan McIlroy <lmcilroy@redhat.com>
Date: Mon, 06 May 2013 03:10:00 +0000
Subject: ext4: limit group search loop for non-extent files

In the case where we are allocating for a non-extent file,
we must limit the groups we allocate from to those below
2^32 blocks, and ext4_mb_regular_allocator() attempts to
do this initially by putting a cap on ngroups for the
subsequent search loop.

However, the initial target group comes in from the
allocation context (ac), and it may already be beyond
the artificially limited ngroups.  In this case,
the limit

	if (group == ngroups)
		group = 0;

at the top of the loop is never true, and the loop will
run away.

Catch this case inside the loop and reset the search to
start at group 0.

[sandeen@redhat.com: add commit msg & comments]

Signed-off-by: Lachlan McIlroy <lmcilroy@redhat.com>
Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Cc: stable@vger.kernel.org
Signed-off-by: Jarod Wilson <jarod@redhat.com>