nfs: Fix matching of the stateids when returning a delegation
Patchwork-id: 58833
O-Subject: [RHEL6.5 PATCH 01/22] NFSv4.1: Fix matching of the stateids when returning a delegation
Bugzilla: 923939
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

nfs41_validate_delegation_stateid is broken if we supply a stateid with
a non-zero sequence id. Instead of trying to match the sequence id,
the function assumes that we always want to error. While this is
true for a delegation callback, it is not true in general.

Also fix a typo in nfs4_callback_recall.

Reported-by: Andy Adamson <andros@netapp.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: 8e663f0e5fabf57065aed1cfdaff5b13057dce23