nfs: The stateid must remain the same for replayed RPC calls
Patchwork-id: 58840
O-Subject: [RHEL6.5 PATCH 10/22] NFSv4: The stateid must remain the same for replayed RPC calls
Bugzilla: 923939
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

If we replay a READ or WRITE call, we should not be changing the
stateid. Currently, we may end up doing so, because the stateid
is only selected at xdr encode time.

This patch ensures that we select the stateid after we get an NFSv4.1
session slot, and that we keep that same stateid across retries.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: 9b20614988199fb03580b335a28250922e902098