propagate proper return codes from blk_get_request callers
Patchwork-id: 58864
O-Subject: [RHEL6 PATCH 2/2 v2] block: propagate proper return codes from blk_get_request callers
Bugzilla: 927918
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

When blk_get_request fails and gfp_flags include __GFP_WAIT, it means
that the queue is dead.  Some call sites misinterpreted this to mean
-ENOMEM, which would then propagate up the call chain (often times
all the way up to user space).  This patch fixes up the return codes.

This is different from the upstream resolution to the problem, which
modifies the call signature of blk_get_request to return an ERROR_PTR.
We can't do that, as it would break kabi.

The upstream discussion on this topic can be found here:
  http://thread.gmane.org/gmane.linux.scsi/80934
and the patches for upstream have not yet been accepted, pending further
work by Joe Lawrence in the scsi_dh code.

This addresses bug 927918.  Comments, as always, are welcome.