gfs2: Use ->dirty_inode()
Patchwork-id: 58866
O-Subject: [RHEL6.5 PATCH][bz887341][TRY #2] GFS2: Use ->dirty_inode()
Bugzilla: 887341
RH-Acked-by: Abhijith Das <adas@redhat.com>

This patch is for bugzilla bug #887341.
This is a combination of these three upstream patches:

GFS2: Use ->dirty_inode()
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2?id=ab9bbda0204dfd0e5342562d9979d1241b14ea5f

GFS2: Don't call file_accessed() with a shared glock
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2?id=3d1626889a64bd5a661544d582036a0a02104a60

GFS2: Use dirty_inode in gfs2_dir_add
http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2?id=343cd8f0d78515da38e41e9351f5ba306cdec84a

It was tested on the afc-east cluster (pats, jets, bills, dolphins)
using the two-hour+ SAS calibration workload.

Original patch description:

The aim of this patch is to use the newly enhanced ->dirty_inode()
super block operation to deal with atime updates, rather than
piggy backing that code into ->write_inode() as is currently
done.

The net result is a simplification of the code in various places
and a reduction of the number of gfs2_dinode_out() calls since
this is now implied by ->dirty_inode().

Some of the mark_inode_dirty() calls have been moved under glocks
in order to take advantage of then being able to avoid locking in
->dirty_inode() when we already have suitable locks.

One consequence is that generic_write_end() now correctly deals
with file size updates, so that we do not need a separate check
for that afterwards. This also, indirectly, means that fdatasync
should work correctly on GFS2 - the current code always syncs the
metadata whether it needs to or not.

Has survived testing with postmark (with and without atime) and
also fsx.

rhbz#887341

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>