just do the necessary things in early register on SoCs
Patchwork-id: 58947
O-Subject: [RHEL 6.5 PATCH 077/176] bcma: just do the necessary things in early register on SoCs
Bugzilla: 757944
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: 757944
Brew: <https://brewweb.devel.redhat.com/taskinfo?taskID=5738553>
Upstream Status: 49655bb8a51565f0375a4f783334c9de78134be5
Tested: By me. Acceptable. See BZ for details

Some parts of the initialization for chip common and the pcie core are
accessing the sprom struct, but it is not initialized at that stage.
Just do the necessary thing in the early register on SoCs and not the
complete initialization to read out the nvram from the flash chip.
After it is possible to read out the nvram, the sprom should be parsed
from it and the full initialization of the cores should be run.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: John W. Linville <linville@tuxdriver.com>