use fallback sprom if sprom on card was not valid
Patchwork-id: 58952
O-Subject: [RHEL 6.5 PATCH 084/176] bcma: use fallback sprom if sprom on card was not valid
Bugzilla: 757944
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: 757944
Brew: <https://brewweb.devel.redhat.com/taskinfo?taskID=5738553>
Upstream Status: 017c4c3bdeb4e6c38094eb2330a57370f68f4853
Tested: By me. Acceptable. See BZ for details

Sometimes the PCIe card indicates that it has a sprom somewhere and we
are able to read the memory region, but it is empty and not valid. In
these cases we should try to use the fallback sprom as a last chance.

This is the case for the PCIe cards in my ASUS RT-N66U (BCM4706 + 2
times BCM4331) and I have heard of someone having the same problem with
an other PCIe card connected to an other Broadcom SoC.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: John W. Linville <linville@tuxdriver.com>