do not initialize deactivated PCIe cores
Patchwork-id: 58961
O-Subject: [RHEL 6.5 PATCH 086/176] bcma: do not initialize deactivated PCIe cores
Bugzilla: 757944
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: 757944
Brew: <https://brewweb.devel.redhat.com/taskinfo?taskID=5738553>
Upstream Status: 2b4766c3008aa99a60415a6f321ac0dad94852f1
Tested: By me. Acceptable. See BZ for details

Before it was tried to initialize the deactivated PCIe core in client
mode, but this causes the SoC to hang. Just do not initialize it at all
and ignore the core it is not working and nothing is connected to it
when the specific bit is set in the boardflags.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: John W. Linville <linville@tuxdriver.com>