perf: Fix finding overflowed PMC in interrupt
Patchwork-id: 59147
O-Subject: [PATCH RHEL6.5 BZ905138 1/2 v3] powerpc/perf: Fix finding overflowed PMC in interrupt
Bugzilla: 905138
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=905138

Description:
------------
If a PMC is about to overflow on a counter that's on an active perf event
(ie. less than 256 from the end) and a _different_ PMC overflows just at this
time (a PMC that's not on an active perf event), we currently mark the event as
found, but in reality it's not as it's likely the other PMC that caused the
IRQ.  Since we mark it as found the second catch all for overflows doesn't run,
and we don't reset the overflowing PMC ever.  Hence we keep hitting that same
PMC IRQ over and over and don't reset the actual overflowing counter.

This is a rewrite of the perf interrupt handler for book3s to get around this.
We now check to see if any of the PMCs have actually overflowed (ie >=
0x80000000).  If yes, record it for active counters and just reset it for
inactive counters.  If it's not overflowed, then we check to see if it's one of
the buggy power7 counters and if it is, record it and continue.  If none of the
PMCs match this, then we make note that we couldn't find the PMC that caused
the IRQ.

Signed-off-by: Michael Neuling <mikey neuling org>
Reviewed-by: Sukadev Bhattiprolu <sukadev linux vnet ibm com>
cc: Paul Mackerras <paulus samba org>
cc: Anton Blanchard <anton samba org>
cc: Linux PPC dev <linuxppc-dev ozlabs org>
Signed-off-by: Benjamin Herrenschmidt <benh kernel crashing org>

v2 - fix whitespace issues pointed out by Stefan.
v3 - fix up the number of parameters that record_and_restart() has after Jiro's perf rebase

RHEL Version Found:
-------------------
RHEL 6.4

Brew:
-----
https://brewweb.devel.redhat.com/taskinfo?taskID=5784825

Upstream:
---------
http://git.kernel.org/?p=linux/kernel/git/benh/powerpc.git;a=commitdiff;h=bc09c219b2e6f9436d06a1a3a10eff97faab371c

Test Status:
------------
Maynard Johnson helped me test this change.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>