kvm: Add x86_hyper_kvm to complete detect_hypervisor_platform check
Patchwork-id: 59150
O-Subject: [patch 1/3] KVM: Add x86_hyper_kvm to complete detect_hypervisor_platform check
Bugzilla: 905688
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

BZ: 905688

commit fc73373b33f5f965f2f82bfbc40ef8e6072e986d upstream
Author: Prarit Bhargava <prarit@redhat.com>

While debugging I noticed that unlike all the other hypervisor code in the
kernel, kvm does not have an entry for x86_hyper which is used in
detect_hypervisor_platform() which results in a nice printk in the
syslog.  This is only really a stub function but it
does make kvm more consistent with the other hypervisors.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: Avi Kivity <avi@redhat.com>
Cc: Gleb Natapov <gleb@redhat.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Cc: Marcelo Tostatti <mtosatti@redhat.com>
Cc: kvm@vger.kernel.org
Signed-off-by: Avi Kivity <avi@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>