perf: Fix perf_event_for_each() to use sibling
Patchwork-id: 59588
O-Subject: [PATCH RHEL6.5 BZ967933 7/7] perf: Fix perf_event_for_each() to use sibling
Bugzilla: 967933
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Michael Ellerman <michael@ellerman.id.au>

Bugzilla: 967933
https://bugzilla.redhat.com/show_bug.cgi?id=967933

upstream status
===============
commit 724b6daa13e100067c30cfc4d1ad06629609dc4e
Author: Michael Ellerman <michael@ellerman.id.au>
Date:   Wed Apr 11 11:54:13 2012 +1000

description
===========
In perf_event_for_each() we call a function on an event, and then
iterate over the siblings of the event.

However we don't call the function on the siblings, we call it
repeatedly on the original event - it seems "obvious" that we should
be calling it with sibling as the argument.

It looks like this broke in commit 75f937f24bd9 ("Fix ctx->mutex
vs counter->mutex inversion").

The only effect of the bug is that the PERF_IOC_FLAG_GROUP parameter
to the ioctls doesn't work.

Signed-off-by: Jarod Wilson <jarod@redhat.com>