configs: enable aesni-intel on x86-32
Patchwork-id: 59819
O-Subject: [RHEL6 PATCH 51/51 v2] [redhat] configs: enable aesni-intel on x86-32
Bugzilla: 786968 786972 969522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=786968
https://bugzilla.redhat.com/show_bug.cgi?id=786972
https://bugzilla.redhat.com/show_bug.cgi?id=969522

To date, we've only been shipping it enabled on x86_64, largely because it
hadn't been ported to x86_32 until semi-recently. This patch series
includes the port to x86_32, so we can turn it on now.

v2: patch the right config file this time

Signed-off-by: Jarod Wilson <jarod@redhat.com>