perf/event: No need to create a thread when handling PERF_RECORD_EXIT
Patchwork-id: 59925
O-Subject: [PATCH RHEL6.5 BZ970529 014/146] perf event: No need to create a thread when handling PERF_RECORD_EXIT
Bugzilla: 970529
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 970529
https://bugzilla.redhat.com/show_bug.cgi?id=970529

upstream status
===============
commit f62d3f0f4596f983ec00495d91c8ddb30268d878
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Sat Oct 6 15:44:59 2012 -0300

description
===========
When we were processing a PERF_RECORD_EXIT event we first used
machine__findnew_thread for both the thread exiting and for its parent,
only to use just the thread struct associated with the one exiting, and
to just delete it.

If it existed, i.e. not created at this very moment in
machine__findnew_thread, it will be moved to the machine->dead_threads
linked list, because we may have hist_entries pointing to it, but if it
was created just do be deleted, it will just sit there with no
references at all.

Use the new machine__find_thread() method so that if it is not there, we
don't create it.

As a bonus the parent thread will also not be created at this point.

Create process_fork() and process_exit() helpers to use this and make
the builtins use it instead of the generic process_task(), ditched by
this patch.