perf: Try to build Documentation when installing
Patchwork-id: 59953
O-Subject: [PATCH RHEL6.5 BZ970529 036/146] perf tools: Try to build Documentation when installing
Bugzilla: 970529
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Borislav Petkov <borislav.petkov@amd.com>

Bugzilla: 970529
https://bugzilla.redhat.com/show_bug.cgi?id=970529

upstream status
===============
commit af3df2cf17f5df079189c3cc88870d28e219496b
Author: Borislav Petkov <borislav.petkov@amd.com>
Date:   Fri Sep 28 19:47:07 2012 +0200

description
===========
There's a portion in the "perf list" output refering to the exact
specification of raw hardware events.

Since this description is in the perf-list manpage, try to build and
install the man pages, warning the user when that is not possible
due to missing packages (xmlto and asciidoc).