perf/hists: Introduce hists__link
Patchwork-id: 60022
O-Subject: [PATCH RHEL6.5 BZ970529 110/146] perf hists: Introduce hists__link
Bugzilla: 970529
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 970529
https://bugzilla.redhat.com/show_bug.cgi?id=970529

upstream status
===============
commit 494d70a18137d18f0728fab7ad4f56aba29d1982
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Nov 8 18:03:09 2012 -0300

description
===========
That given two hists will find the hist_entries (buckets) in the second
hists that are for the same bucket in the first and link them, then it
will look for all buckets in the second that don't have a counterpart in
the first and will create a dummy counterpart that will then be linked
to the entry in the second.

For multiple events this will be done pairing the leader with all the
other events in the group, so that in the end the leader will have all
the buckets in all the hists in a group, dummy or not while the other
hists will be left untouched.