perf/annotate: Don't try to follow jump target on PLT symbols
Patchwork-id: 60026
O-Subject: [PATCH RHEL6.5 BZ970529 115/146] perf annotate: Don't try to follow jump target on PLT symbols
Bugzilla: 970529
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Namhyung Kim <namhyung.kim@lge.com>

Bugzilla: 970529
https://bugzilla.redhat.com/show_bug.cgi?id=970529

upstream status
===============
commit 32ae1efd9d40645601cd4e09fa83a2711dd1ad6d
Author: Namhyung Kim <namhyung.kim@lge.com>
Date:   Sat Nov 10 02:27:15 2012 +0900

description
===========
The perf annotate browser on TUI can identify a jump target for a
selected instruction.  It assumes that the jump target is within the
function but it's not the case of PLT symbols which have offset out of
the function as a target.

Since it caused a segmentation fault, do not try to follow jump target
on the PLT symbols.