fsache: Limit the number of I/O error reports for a cache
Patchwork-id: 60467
O-Subject: [RHEL6 PATCH 05/22] FS-Cache: Limit the number of I/O error reports for a cache [BZ 921178]
Bugzilla: 921178
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Limit the number of I/O error reports for a cache to 1 to prevent massive
amounts of noise.  After the first I/O error the cache is taken off line
automatically, so must be restarted to resume caching.

Signed-off-by: David Howells <dhowells@redhat.com>
Upstream-commit: 75bc411388f4aeb9fb0381bd56eb5d67193ed9a1
Bugzilla: 921178