fscache: remove spin_lock() from the condition in while()
Patchwork-id: 60475
O-Subject: [RHEL6 PATCH 13/22] fs/fscache: remove spin_lock() from the condition in while()
Bugzilla: 921178
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Sebastian Andrzej Siewior <bigeasy@linutronix.de>

The spinlock() within the condition in while() will cause a compile error
if it is not a function. This is not a problem on mainline but it does not
look pretty and there is no reason to do it that way.
That patch writes it a little differently and avoids the double condition.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: David Howells <dhowells@redhat.com>